requirements.txt
asgiref==3.5.2
brotlipy @ file:///Users/runner/miniforge3/conda-bld/brotlipy_1648854242877/work
certifi==2022.5.18.1
cffi @ file:///Users/runner/miniforge3/conda-bld/cffi_1625835350093/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1644853463426/work
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1602866480661/work
conda==4.13.0
conda-package-handling @ file:///Users/runner/miniforge3/conda-bld/conda-package-handling_1649385125392/work
cryptography @ file:///Users/runner/miniforge3/conda-bld/cryptography_1652967108255/work
Django==4.0.4
django-classy-tags==3.0.1
django-taggit==3.0.0
django-taggit-templatetags2==1.6.1
djangorestframework==3.13.1
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1642433548627/work
pycosat @ file:///Users/runner/miniforge3/conda-bld/pycosat_1649384941891/work
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1636257122734/work
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1643496850550/work
PySocks @ file:///Users/runner/miniforge3/conda-bld/pysocks_1648857374584/work
pytz==2022.1
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1641580202195/work
ruamel-yaml-conda @ file:///Users/runner/miniforge3/conda-bld/ruamel_yaml_1653464548430/work
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
sqlparse==0.4.2
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1649051611147/work
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1647489083693/work
